pyproject.toml
[project]
name = "ragproductionapp"
version = "0.1.0"
description = "RAG PDF Chat Application"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.116.1",
    "groq>=1.0.0",
    "inngest>=0.5.6",
    "llama-index-core>=0.14.0",
    "llama-index-readers-file>=0.5.4",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.49.1",
    "uvicorn>=0.35.0",
]